+++ frontend/src/pages/Statistics/StatisticsPage.js
@@ -99,7 +99,7 @@ class StatisticsPage extends Component {
                 <div className={styles.container}>
                     <div className={styles.eventCard}>
                         <h3>
-                            Total d'événements en {this.getCurrentMonthName()} :{" "}
+                            Total d'évènements en {this.getCurrentMonthName()} :{" "}
                             {this.state.events?.totalEvents}
                         </h3>
                         <p className={styles.labelStats}>