+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 -   Add events by employee stats card
 -   Add download document link
 -   Add tel and mail link to table clients
+-   Add reset filter to clients and calendar pages
 
 ## Fixes